requirements.txt
dj-database-url==0.5.0
Django==3.0.8
django-extensions==3.0.2
djangorestframework==3.11.0
psycopg2-binary==2.8.5
python-decouple==3.3
requests==2.24.0